커뮤니티 삭제 확인 알림창 추가

+++ multi_final/front/workspace_react/src/page/communityCommunication/CommunityReply.js
@@ -17,11 +17,11 @@ const CommunityReply = () => {
   const[contentCnt, setContentCnt] = useState(0);
   
   function Delete(num){
-       
-    axios.delete(`http://localhost:8085/deleteRep/${num}`)
+    if(window.confirm("댓글을 삭제하시겠습니까?")){
+      axios.delete(`http://localhost:8085/deleteRep/${num}`)
          .then(window.location='/communication/'+no).catch(err=>console.log(err))
-
-      }
+    }
+  }
   
 
       const [rg_name,setrg_name]=useState(''); //댓글쓴 이름
